global working

+++ GoBi/src/blatt1/AlignmentMax.java
@@ -7,21 +7,21 @@ package blatt1;
 public class AlignmentMax {
     
     private int i,j;
-    private int max;
+    private double max;
     private final String mode;
     
-    public AlignmentMax(int i, int j, int max, String mode){
+    public AlignmentMax(int i, int j, double max, String mode){
         this.i = i;
         this.j = j;
         this.max = max;
         this.mode = mode;
     }
     
-    public int getMax(){
+    public double getMax(){
         return max;
     }
     
-    public void setMax(int i, int j, int max){
+    public void setMax(int i, int j, double max){
         this.i = i;
         this.j = j;
         this.max = max;
